fix:pool state処理修正
